+++ src/cpu/opcodes_cb.rs
@@ -65,7 +65,7 @@ impl CPU {
     }
 
 
-    fn run_operation_on_index<F>(&mut self, memory: &mut Memory, index: u8, mut operation: F) -> u64
+    fn run_operation_on_index<F>(&mut self, memory: &mut Memory, index: u8, mut operation: F) -> u16
     where
         F: FnMut(&mut Self, u8) -> u8,
     {
@@ -83,7 +83,7 @@ impl CPU {
         }
     }
 
-    fn run_res_set_operation_on_index<F>(&mut self, memory: &mut Memory, index: u8, bit: u8, mut operation: F, hl_cycles: u64) -> u64
+    fn run_res_set_operation_on_index<F>(&mut self, memory: &mut Memory, index: u8, bit: u8, mut operation: F, hl_cycles: u16) -> u16
     where
         F: FnMut(&mut Self, u8, u8) -> u8,
     {
@@ -101,7 +101,7 @@ impl CPU {
         }
     }
 
-    fn run_bit_operation_on_index<F>(&mut self, memory: &mut Memory, index: u8, bit: u8, mut operation: F, hl_cycles: u64) -> u64
+    fn run_bit_operation_on_index<F>(&mut self, memory: &mut Memory, index: u8, bit: u8, mut operation: F, hl_cycles: u16) -> u16
     where
         F: FnMut(&mut Self, u8, u8),
     {
